+++ src/static/test.js
@@ -1,7 +1,9 @@
 const sources = document.getElementById("sources");
 const campaigns = document.getElementById("campaigns");
 const agents = document.getElementById("agents");
-const type = document.getElementById("type");
+const types = document.getElementById("types");
+const ingroups = document.getElementById("ingroups");
+const dispos = document.getElementById("dispos");
 
 const getSources = () => {
     axios.get(SCRIPT_ROOT + "api/_get_sources")
@@ -52,7 +54,7 @@ const getTypes = () => {
                 let option = document.createElement("option");
                 option.value = elem;
                 option.text = response.data[elem]
-                type.appendChild(option);
+                types.appendChild(option);
             };           
         });
 }
@@ -61,7 +63,7 @@ const getCampaignsByType = () => {
     while(campaigns.options.length > 0) {
         campaigns.remove(0)    
     }
-    axios.get(SCRIPT_ROOT + "api/_get_campaigns_by_type/" + sources.value + "/" + type.value)
+    axios.get(SCRIPT_ROOT + "api/_get_campaigns_by_type/" + sources.value + "/" + types.value)
         .then( response  => {
             response.data.map(elem => {
                 let option = document.createElement("option");
@@ -69,6 +71,36 @@ const getCampaignsByType = () => {
                 option.text = elem[1];
                 campaigns.appendChild(option);
             }); 
+        });
+}
+
+const getIngroupsByCampaign = () => {
+    while(ingroups.options.length > 0) {
+        ingroups.remove(0)
+    }
+    axios.get(SCRIPT_ROOT + "api/_get_ingroups_by_campaign/" + sources.value + "/" + campaigns.value)
+        .then( response => {
+            response.data.map(elem => {
+                let option = document.createElement("option");
+                option.value = elem[0];
+                option.text = elem[1];
+                ingroups.appendChild(option);
+            });
+        });
+}
+
+const getDisposByCampaign = () => {
+    while(dispos.options.length > 0) {
+        dispos.remove(0)
+    }
+    axios.get(SCRIPT_ROOT + "api/_get_dispos_by_campaign/" + sources.value + "/" + campaigns.value)
+        .then( response => {
+            response.data.map(elem => {
+                let option = document.createElement("option");
+                option.value = elem[0];
+                option.text = elem[1];
+                dispos.appendChild(option);
+            });
         });
 }
 
@@ -83,8 +115,10 @@ sources.addEventListener("change", () => {
 
 campaigns.addEventListener("change", () => {
     getAgentsByCampaing();
+    getIngroupsByCampaign();
+    getDisposByCampaign();
 });
 
-type.addEventListener("change", () => {
+types.addEventListener("change", () => {
     getCampaignsByType();
 });